Implemented first test version for rough update execution in moodleUpdateCheck.sh

+++ moodleUpdateCheck.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -e
-#image_version="$APP_VERSION"
-image_version="5.0.0"
+image_version="$APP_VERSION"
+#image_version="5.0.0"#Provoke update
 # checks if image version(new) is greater than current installed version
 version_greater() {
 	if [[ $1 = $2 ]]; then echo "Already up to date"; return 0;
@@ -9,6 +9,11 @@ version_greater() {
     elif [[ $1 < $2 ]]; then echo "Initializing Moodle $image_version ..."; return 1;
 	else echo "Unexpected behaviour, exiting version check"; return 0; fi
 }
+
+if [ -f /bitnami/moodledata/UpdateFailed ]; then
+    echo "=== UpdateFailed file exists, please resolve the image problem manually ==="
+    /opt/bitnami/scripts/moodle/entrypoint.sh "/opt/bitnami/scripts/moodle/run.sh"
+fi
 
 #Get the current installed version
 installed_version="0.0.0"
@@ -25,37 +30,40 @@ else
     exit 0;
 fi
 
+#Is needed to check for success inside the container at the end
+pre_update_version=installed_version;
 echo "The new Moodle Image version is $APP_VERSION";
 
 if version_greater "$installed_version" "$image_version";
 then
-echo "Skipping Upgrade process"; exit 0
-#TODO end the script properly to launch the container successfully
+echo "=== Skipping Update process and starting Moodle ===";
+/opt/bitnami/scripts/moodle/entrypoint.sh "/opt/bitnami/scripts/moodle/run.sh"
 else
-    echo "=== Preparing Update ==="
-    #apt-get install curl gnupg
-    # apt-get install apt-transport-https --yes
-    # echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/helm.gpg] https://baltocdn.com/helm/stable/debian/ all main" | tee /etc/apt/sources.list.d/helm-stable-debian.list
-    # apt-get update
-    # apt-get -y install helm
-    # helm repo add bitnami https://charts.bitnami.com/bitnami
 
-    echo "=== Starting Update ==="
-
-    echo "=== Enable Maintenance Mode ==="
-    echo '<h1>Sorry, maintenance in progress</h1>' > /bitnami/moodledata/climaintenance.html
-
-    echo "=== Turn off liveness and readiness probe ==="
-    helm upgrade --reuse-values --set livenessProbe.enabled=false --set readinessProbe.enabled=false moodle  bitnami/moodle --version {{ moodle_chart_version }} --namespace {{ moodle_namespace }}
-    #Helm command not found
-    echo "=== Taking a Backup ==="
-    if [ -d "/bitnami/moodledata/moodle-backup" ]; then
-        rm -r /bitnami/moodledata/moodle-backup
+    if ! [ -a /bitnami/moodledata/climaintenance.html ]; then
+        echo "=== Enable Maintenance Mode ==="
+        echo '<h1>Sorry, maintenance in progress</h1>' > /bitnami/moodledata/climaintenance.html
+        sleep 5
+        #The backup is only done once in the first run so we don't accidentally overwrite it
+        echo "=== Taking a Backup ===" 
+        if [ -d "/bitnami/moodledata/moodle-backup" ]; then
+            rm -r /bitnami/moodledata/moodle-backup
+        fi
+        mkdir -p /bitnami/moodledata/moodle-backup
+        cp -rp /bitnami/moodle/* /bitnami/moodledata/moodle-backup
     fi
-    mkdir -p /bitnami/moodledata/moodle-backup
-    cp -rp /bitnami/moodle/* /bitnami/moodledata/moodle-backup
 
-    # TODO include correct way for Plugins
+    if ! [ -a /bitnami/moodledata/CliUpdate ]; then
+        echo "=== Create required Files for Update ==="
+        touch /bitnami/moodledata/CliUpdate
+        sleep 20 #Ensure sufficient time for possible full update
+    fi
+
+    echo "=== Starting new Moodle Download ==="
+
+    #echo "=== Turn off liveness and readiness probe ==="
+    #helm upgrade --reuse-values --set livenessProbe.enabled=false --set readinessProbe.enabled=false moodle  bitnami/moodle --namespace {{ moodle_namespace }}
+    
     curl {{ moodle_update_url }} -o /tmp/moodle.tgz
 
     echo "=== Download complete ==="
@@ -84,8 +92,37 @@ else
     # chown -R 1001:root /bitnami/moodle/*
 
     # echo "=== Turn liveness probe back on again ==="
-    helm upgrade --reuse-values --set livenessProbe.enabled=true --set readinessprobe.enable=true moodle bitnami/moodle --version {{ moodle_chart_version }} --namespace {{ moodle_namespace }}
+    #helm upgrade --reuse-values --set livenessProbe.enabled=true --set readinessprobe.enable=true moodle bitnami/moodle --version {{ moodle_chart_version }} --namespace {{ moodle_namespace }}
 
-    # echo "=== Disable Maintenance Mode ==="
-    rm  /bitnami/moodle-data/climaintenance.html
+    #If success
+    #Get the new installed version
+    post_update_version="0.0.0"
+    if [ -f /bitnami/moodle/version.php ]; then
+        LINE=$(grep release /bitnami/moodle/version.php)
+        REGEX="release\s*=\s*'([0-9]+\.[0-9]+\.[0-9])"
+        if [[ $LINE =~ $REGEX ]]; then
+            echo "Installed Moodle version:" ${BASH_REMATCH[1]}
+            post_update_version=${BASH_REMATCH[1]}
+        fi
+    else
+        #TODO What happens if there is no Moodle installed?
+        echo "=== Update failed, no Moodle Version detected ==="
+        exit 0;
+    fi
+
+    if [ $post_update_version == $APP_VERSION ]; then
+        echo "=== Update to new Version $post_update_version successful ==="
+        #Cleanup
+        echo "=== Disable Maintenance Mode ==="
+        rm  /bitnami/moodledata/climaintenance.html
+        rm  /bitnami/moodledata/CliUpdate
+        #rm -r /bitnami/moodledata/moodle-backup
+        #rm -r /bitnami/moodledata/updated-moodle
+        echo "=== Starting new Moodle version ==="
+        /opt/bitnami/scripts/moodle/entrypoint.sh "/opt/bitnami/scripts/moodle/run.sh"
+    elif [ $post_update_version == $pre_update_version ]; then
+        echo "=== Update failed, old Version still installed ===" #Do we want to keep running until manual intervention?
+        exit 0;
+    fi
+
 fi
